+++ FamTec/FamTec/FamTec/Server/Controllers/User/UserController.cs
@@ -27,23 +27,18 @@ namespace FamTec.Server.Controllers.User
         [Route("sign/GetPlaceUsers")]
         public async ValueTask<IActionResult> GetUserList()
         {
-            ResponseList<ListUser>? model = await UserService.GetPlaceUserList(HttpContext);
-            
-            if (model is not null)
-            {
-                if (model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
-                    return BadRequest();
-                }
-            }
-            else
-            {
+            if (HttpContext is null)
                 return BadRequest();
-            }
+
+            ResponseList<ListUser>? model = await UserService.GetPlaceUserList(HttpContext);
+
+            if (model is null)
+                return BadRequest();
+
+            if (model.code == 200)
+                return Ok(model);
+            else
+                return BadRequest();
         }
 
         /// <summary>
@@ -57,23 +52,27 @@ namespace FamTec.Server.Controllers.User
         [Route("sign/AddUser")]
         public async ValueTask<IActionResult> AddUser([FromForm]UsersDTO dto, [FromForm]IFormFile? files)
         {
+            if (HttpContext is null)
+                return BadRequest();
+
+            // 파일크기 1MB 지정
+            if (files is not null)
+            {
+                if (files.Length > 1048576)
+                {
+                    return Ok(new ResponseUnit<UsersDTO>() { message = "이미지 업로드는 1MB 이하만 가능합니다.", data = null, code = 200 });
+                }
+            }
+
             ResponseUnit<UsersDTO>? model = await UserService.AddUserService(HttpContext, dto, files);
 
-            if(model is not null)
-            {
-                if(model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
-                    return BadRequest();
-                }
-            }
-            else
-            {
+            if(model is null)
                 return BadRequest();
-            }
+
+            if (model.code == 200)
+                return Ok(model);
+            else
+                return BadRequest();
         }
 
         [AllowAnonymous]
@@ -81,22 +80,17 @@ namespace FamTec.Server.Controllers.User
         [Route("sign/DetailUser")]
         public async ValueTask<IActionResult> DetailUser([FromQuery]int? id)
         {
-            ResponseUnit<UsersDTO> model = await UserService.GetUserDetails(HttpContext, id);
-            if(model is not null)
-            {
-                if(model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
-                    return BadRequest();
-                }
-            }
-            else
-            {
+            if (HttpContext is null)
                 return BadRequest();
-            }
+
+            ResponseUnit<UsersDTO> model = await UserService.GetUserDetails(HttpContext, id);
+            if(model is null)
+                return BadRequest();
+
+            if (model.code == 200)
+                return Ok(model);
+            else
+                return BadRequest();
         }
 
         [AllowAnonymous]
@@ -104,23 +98,18 @@ namespace FamTec.Server.Controllers.User
         [Route("sign/DeleteUser")]
         public async ValueTask<IActionResult> DeleteUser([FromQuery] List<int> delIdx)
         {
+            if (HttpContext is null)
+                return BadRequest();
+
             ResponseUnit<int?> model = await UserService.DeleteUserService(HttpContext, delIdx);
 
-            if(model is not null)
-            {
-                if(model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
-                    return BadRequest();
-                }
-            }
-            else
-            {
+            if(model is null)
                 return BadRequest();
-            }
+
+            if (model.code == 200)
+                return Ok(model);
+            else
+                return BadRequest();
         }
 
         [AllowAnonymous]
@@ -128,42 +117,41 @@ namespace FamTec.Server.Controllers.User
         [Route("sign/UpdateUser")]
         public async ValueTask<IActionResult> UpdateUser([FromForm]UpdateUserDTO dto, [FromForm]IFormFile? files)
         {
+            if (HttpContext is null)
+                return BadRequest();
+
             ResponseUnit<UpdateUserDTO>? model = await UserService.UpdateUserService(HttpContext, dto, files);
 
-            if(model is not null)
-            {
-                if(model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
-                    return BadRequest();
-                }
-            }
-            else
-            {
+            if(model is null)
                 return BadRequest();
-            }
+
+            if (model.code == 200)
+                return Ok(model);
+            else
+                return BadRequest();
 
         }
 
+        /// <summary>
+        /// 엑셀 IMPORT
+        /// </summary>
+        /// <param name="file"></param>
+        /// <returns></returns>
         [AllowAnonymous]
         [HttpPost]
         [Route("sign/ImportUser")]
         public async Task<IActionResult> UploadFile([FromForm] IFormFile? file)
         {
+            if (HttpContext is null)
+                return BadRequest();
+
             ResponseUnit<string>? model = await UserService.ImportUserService(HttpContext, file);
-            if(model is not null)
+            if(model is null)
+                return BadRequest();
+
+            if (model.code == 200)
             {
-                if(model.code == 200)
-                {
-                    return Ok(model);
-                }
-                else
-                {
-                    return BadRequest();
-                }
+                return Ok(model);
             }
             else
             {
